+++ src/test/java/de/caritas/cob/agencyservice/api/admin/service/DioceseAdminServiceTenantSupportIT.java
@@ -1,20 +1,41 @@
 package de.caritas.cob.agencyservice.api.admin.service;
 
 import de.caritas.cob.agencyservice.AgencyServiceApplication;
+import de.caritas.cob.agencyservice.api.tenant.TenantContext;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.jdbc.Sql;
+import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.transaction.annotation.Transactional;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = AgencyServiceApplication.class)
 @TestPropertySource(properties = "spring.profiles.active=testing")
 @AutoConfigureTestDatabase(replace = Replace.ANY)
-public class DioceseAdminServiceIT extends DioceseAdminServiceITBase {
+@DirtiesContext(classMode = ClassMode.BEFORE_CLASS)
+@TestPropertySource(properties = "multitenancy.enabled=true")
+@Transactional
+@Sql(value = "/setTenants.sql", executionPhase = ExecutionPhase.BEFORE_TEST_METHOD)
+public class DioceseAdminServiceTenantSupportIT extends DioceseAdminServiceITBase {
+
+  @Before
+  public void beforeEach(){
+    TenantContext.setCurrentTenant("1");
+  }
+
+  @After
+  public void afterEach(){
+    TenantContext.clear();
+  }
 
   @Test
   public void findAllDioceses_Should_returnOneResult_When_perPageIsSetToOne() {
